+++ NuRadioReco/utilities/__init__.py
@@ -1,0 +1,1 @@
+from NuRadioReco.utilities._deprecated import bandpass_filter
